pyproject.toml
[project]
name = "telegram-report-bot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "aiosqlite>=0.22.1",
    "beautifulsoup4>=4.14.3",
    "chardet>=7.0.1",
    "finvizfinance>=1.3.0",
    "google-generativeai>=0.8.6",
    "loguru>=0.7.3",
    "lxml>=6.0.2",
    "openpyxl>=3.1.5",
    "oracledb>=3.4.2",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.6",
    "pytz>=2026.1.post1",
    "requests>=2.32.5",
    "selenium>=4.41.0",
    "webdriver-manager>=4.0.2",
    "wget>=3.2",
]
